+++ packages/components/src/Banner/Banner.tsx
@@ -1,7 +1,7 @@
 import Heading, { HeadingElement } from "../Heading";
 import Button from "../Button";
 import CloseIcon from "../Icons/Close";
-import NotificationIcon from "../NotificationIcon";
+import NotificationIcon from "../common/NotificationIcon";
 import React from "react";
 import Text from "../Text";
 import clsx from "clsx";
