+++ packages/components/src/Toast/Toast.tsx
@@ -1,6 +1,6 @@
 import Button from "../Button";
 import CloseIcon from "../Icons/Close";
-import NotificationIcon from "../NotificationIcon";
+import NotificationIcon from "../common/NotificationIcon";
 import React from "react";
 import Text from "../Text";
 import clsx from "clsx";
